+++ skops/io/_general.py
@@ -509,13 +509,15 @@ def method_get_state(obj: Any, save_context: SaveContext) -> dict[str, Any]:
     # dependent on a specific instance of an object.
     # It stores the state of the object the method is bound to,
     # and prepares both to be persisted.
+    owner = obj.__self__
+    func_name = obj.__func__.__name__
     res = {
-        "__class__": obj.__class__.__name__,
+        "__class__": owner.__class__.__name__,
         "__module__": get_module(obj),
         "__loader__": "MethodNode",
         "content": {
-            "func": obj.__func__.__name__,
-            "obj": get_state(obj.__self__, save_context),
+            "func": func_name,
+            "obj": get_state(owner, save_context),
         },
     }
     return res
@@ -529,12 +531,31 @@ class MethodNode(Node):
         trusted: Optional[Sequence[str]] = None,
     ) -> None:
         super().__init__(state, load_context, trusted)
+        obj = get_tree(state["content"]["obj"], load_context, trusted=trusted)
+        if self.module_name != obj.module_name or self.class_name != obj.class_name:
+            raise ValueError(
+                f"Expected object of type {self.module_name}.{self.class_name}, got"
+                f" {obj.module_name}.{obj.class_name}. This is probably due to a"
+                " corrupted or a malicious file."
+            )
         self.children = {
-            "obj": get_tree(state["content"]["obj"], load_context, trusted=trusted),
+            "obj": obj,
             "func": state["content"]["func"],
         }
         # TODO: what do we trust?
         self.trusted = self._get_trusted(trusted, [])
+
+    def get_unsafe_set(self) -> set[str]:
+        res = super().get_unsafe_set()
+        obj_node = self.children["obj"]
+        res.add(
+            obj_node.module_name  # type: ignore
+            + "."
+            + obj_node.class_name  # type: ignore
+            + "."
+            + self.children["func"]
+        )
+        return res
 
     def _construct(self):
         loaded_obj = self.children["obj"].construct()
@@ -658,6 +679,11 @@ class OperatorFuncNode(Node):
         trusted: Optional[Sequence[str]] = None,
     ) -> None:
         super().__init__(state, load_context, trusted)
+        if self.module_name != "operator":
+            raise ValueError(
+                f"Expected module 'operator', got {self.module_name}. This is probably"
+                " due to a corrupted or a malicious file."
+            )
         self.trusted = self._get_trusted(trusted, [])
         self.children["attrs"] = get_tree(state["attrs"], load_context, trusted=trusted)